Use config to include packages
Pulumi did not correctly identify `qs` and `uuid` as being module import
captures in the function.  This led to runtime failures due to the
packages being missing from the lambda code upload.  This is being
tracked by https://github.com/pulumi/pulumi/issues/1588.  In the
meantime, we can use manual configuration to include the packages in
the uploaded functions.